requirements.txt
absl-py==0.13.0
appdirs==1.4.4
astunparse==1.6.3
atpublic==2.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cryptography==3.4.7
decorator==4.4.2
dictdiffer==0.8.1
diskcache==5.2.1
distro==1.5.0
dpath==2.0.1
dulwich==0.20.23
dvc==2.4.1
flatbuffers==1.12
flatten-dict==0.4.0
flufl.lock==3.2
fsspec==2021.6.0
ftfy==6.0.3
funcy==1.16
future==0.18.2
gast==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.30.0
google-api-python-client==2.9.0
google-auth==1.31.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grandalf==0.6
grpcio==1.34.1
h5py==3.1.0
httplib2==0.19.1
idna==2.10
jsonpath-ng==1.5.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
mailchecker==4.0.9
Markdown==3.3.4
nanotime==0.5.2
networkx==2.5.1
numpy==1.19.5
oauth2client==4.1.3
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
pathspec==0.8.1
phonenumbers==8.12.25
Pillow==8.2.0
ply==3.11
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
PyDrive2==1.8.3
pygit2==1.6.1
Pygments==2.9.0
pygtrie==2.3.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-benedict==0.24.0
python-dateutil==2.8.1
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rich==10.4.0
rsa==4.7.2
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
shortuuid==1.0.1
shtab==1.3.6
six==1.15.0
smmap==4.0.0
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
tqdm==4.61.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.5
voluptuous==0.12.1
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
xmltodict==0.12.0
zc.lockfile==2.0
